% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.study.R
\name{chain.study}
\alias{chain.study}
\title{Convergence graphs for the simulated values}
\usage{
chain.study(mcmclist, ...)
}
\arguments{
\item{mcmclist}{A \emph{mcmc.2pnob} or \emph{mcmc.3pnob} class object.}

\item{...}{Further arguments.}
}
\value{
Convergence graphs for the study of the simulated values for an MCMC
marginal chain.
}
\description{
Convergence graphs for the study of the simulated values for an MCMC
marginal chain.
}
\details{
The top left graph displays the sequence of simulated values and the top
right graph displays the lag correlations of the sequence as a function of
the lag value. The bottom left graph is a histogram of the simulated values
and the bottom right graph is the box plot of the simulated values.
}
\references{
Johnson, V. E. & Albert, J. H. (1999). Ordinal Data Modeling.
New York: Springer.
}
\seealso{
\code{\link{mcmc.2pnob}}, \code{\link{mcmc.3pnob}} and
\code{\link{continue.mcmc}}.
}
\author{
Javier Martínez
}
