% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continue.mcmc.R
\name{continue.mcmc.default}
\alias{continue.mcmc.default}
\title{Continue MCMC for the Estimation of the Two-Parameter or
Three-Parameter Normal Ogive Model}
\usage{
\method{continue.mcmc}{default}(mcmclist, ...)
}
\arguments{
\item{mcmclist}{A \emph{bairt} class object (\emph{mcmc.2pnob} or
\emph{mcmc.3pnob}).}

\item{...}{Further arguments.}
}
\value{
An \emph{mcmc.2pnob or mcmc.3pnob} object.
}
\description{
This is a function for \emph{bairt} objects. You can use \emph{continue.mcmc}
for continue the MCMC sampling the Two-Parameter or Three-Parameter
normal ogive item response model.
}
\details{
If any argument (\emph{final.values}, \emph{c.prior}, \emph{iter},
\emph{burning}, \emph{thin} or \emph{parts}) is NULL, \emph{continue.mcmc}
take the value of the \emph{mcmclist}.
}
\references{
Johnson, V. E., & Albert, J. H. (1999). Ordinal Data Modeling.
New York: Springer.

A.A.  Beguin, A, A & Glas, C.A.W. (2001). MCMC Estimation and
Some Model-Fit Analysis of Multidimensional IRT Models. Psychometrika,
66, 541-562.
}
\seealso{
\code{\link{mcmc.2pnob}} and \code{\link{mcmc.3pnob}}.
}
\author{
Javier Martínez
}
