% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irc.R
\name{irc}
\alias{irc}
\title{Plot of posterior density of the item response curve}
\usage{
irc(mcmclist, ...)
}
\arguments{
\item{mcmclist}{A \emph{mcmc.2pnob} or \emph{mcmc.3pnob} class object.}

\item{...}{Further arguments.}
}
\value{
Plot of posterior density of the item response curve for de \emph{j-th}
\code{item}.
}
\description{
Plot of posterior density of the item response curve for the \emph{j-th}
\emph{item}.
}
\details{
The solid line corresponds to the location of the posterior mean and the
points correspond to the percentiles determined for \emph{prob}.
If \emph{prob = c(0.05, 0.95)} is equivalent to the 5th and 95th percentils
of the posterior density.
}
\references{
Johnson, V. E. & Albert, J. H. (1999). Ordinal Data Modeling.
New York: Springer.
}
\seealso{
\code{\link{mcmc.2pnob}}, \code{\link{mcmc.3pnob}} and
\code{\link{continue.mcmc.bairt}}.
}
\author{
Javier Martínez
}
