% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.plot.R
\name{parameter.plot.default}
\alias{parameter.plot.default}
\title{Graph of marginal posterior densities}
\usage{
\method{parameter.plot}{default}(mcmclist, ...)
}
\arguments{
\item{mcmclist}{A \emph{mcmc.2pnob} or \emph{mcmc.3pnob} class object.}

\item{...}{Further arguments.}
}
\value{
Graph of posterior densities of the item parameter (\emph{a}, \emph{b} or
\emph{c}).
}
\description{
Graph of marginal posterior densities for the item parameters (\emph{a},
\emph{b} or \emph{c}).
}
\details{
Graph of marginal posterior densities of the item parameter \emph{a, b}
for \emph{mcmc.2pnob} object or \emph{a, b, c} for \emph{mcmc.3pnob object}.
The center of error bar corresponds to the marginal posterior mean and
the extremes correspond to percentiles of the marginal posterior density
(These are delimited for \emph{prob}). For example,
\emph{prob = c(0.05, 0.95)} is equivalent to the 5th and 95th percentiles
of the marginal posterior density.
}
\references{
Johnson, V. E. & Albert, J. H. (1999). Ordinal Data Modeling.
New York: Springer.
}
\seealso{
\code{\link{mcmc.2pnob}}, \code{\link{mcmc.3pnob}} and
\code{\link{continue.mcmc.bairt}}.
}
\author{
Javier Martínez
}
